Update documentation to new TestListeners
This commit updates the documentation on TestListeners, to include the new test listeners (such as `beforeSpecClass` and `afterSpecClass`), while also adapting the old listeners to the new ones (the old ones are deprecated as of v3.2.0).

A small detail was also updated on the documentation, to include AnnotationSpec in the testing styles.

Fixes #535